Delete old GUI implementation, move new GUI to correct package.
