+++ src/main/java/com/sulphate/chatcolor2/gui/item/impl/ColourItem.java
@@ -1,11 +1,11 @@
-package com.sulphate.chatcolor2.newgui.item.impl;
+package com.sulphate.chatcolor2.gui.item.impl;
 
 import com.sulphate.chatcolor2.data.PlayerData;
-import com.sulphate.chatcolor2.newgui.Gui;
-import com.sulphate.chatcolor2.newgui.item.ItemStackTemplate;
-import com.sulphate.chatcolor2.newgui.item.ComplexGuiItem;
-import com.sulphate.chatcolor2.newgui.item.PermissibleItem;
-import com.sulphate.chatcolor2.newgui.item.SelectableItem;
+import com.sulphate.chatcolor2.gui.Gui;
+import com.sulphate.chatcolor2.gui.item.ItemStackTemplate;
+import com.sulphate.chatcolor2.gui.item.ComplexGuiItem;
+import com.sulphate.chatcolor2.gui.item.PermissibleItem;
+import com.sulphate.chatcolor2.gui.item.SelectableItem;
 import com.sulphate.chatcolor2.utils.InventoryUtils;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
