+++ src/main/java/com/sulphate/chatcolor2/gui/item/impl/SimpleGuiItem.java
@@ -1,7 +1,7 @@
-package com.sulphate.chatcolor2.newgui.item.impl;
+package com.sulphate.chatcolor2.gui.item.impl;
 
-import com.sulphate.chatcolor2.newgui.item.ItemStackTemplate;
-import com.sulphate.chatcolor2.newgui.item.GuiItem;
+import com.sulphate.chatcolor2.gui.item.ItemStackTemplate;
+import com.sulphate.chatcolor2.gui.item.GuiItem;
 import org.bukkit.inventory.ItemStack;
 
 public class SimpleGuiItem extends GuiItem {
