--- src/main/java/com/sulphate/chatcolor2/newgui/item/ItemType.java
@@ -1,8 +1,0 @@
-package com.sulphate.chatcolor2.newgui.item;
-
-public enum ItemType {
-    COLOUR,
-    MODIFIER,
-    INVENTORY,
-    FILLER
-}
